refactor: move scripts to a single scripts.ts file in utils
